pyproject.toml
[tool.poetry]
name = "herre"
version = "0.3.31"
readme = "README.md"
description = "oauth2/openid client tailored to pyqt and async environments"
authors = ["jhnnsrs <jhnnsrs@gmail.com>"]
license = "CC BY-NC 3.0"
packages = [{ include = "herre" }]

[tool.poetry.dependencies]
python = "^3.8"
aiohttp = ">3.7.4"
certifi = ">2021"
oauthlib = "^3.1.1"
pydantic = ">=1.9.0, <2"
koil = ">=0.3.5"
QtPy = { version = "^2.0.1", optional = true }

[tool.poetry.extras]
qt = ["QtPy"]
windowed = ["PyQtWebEngine"]


[tool.mypy]
exclude = ["venv/", "tests/", "examples/"]
ignore_missing_imports = true


[tool.ruff]
extend-select = ["ANN", "D1"]
extend-ignore = ["ANN101", "ANN002", "ANN003", "D100", "ANN401"]

# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
    "tests",
    "examples",
]

# Same as Black.
line-length = 300

[tool.poetry.group.dev.dependencies]
autoflake = "^1.7.7"
pytest = "^7.2.0"
testcontainers = "^3.7.0"
pytest-qt = "^4.2.0"
pytest-asyncio = "^0.20.2"
aiohttp = "^3.8.3"
black = ">=22"
pytest-cov = "^4.0.0"
ruff = "^0.0.282"
mypy = "^1.7.1"
pydoc-markdown = "^4.8.2"
aioresponses = "^0.7.6"
pyqt5-stubs = "^5.15.6.0"


[tool.pytest.ini_options]
qt_api = "pyqt5"
markers = [
    "integration: marks tests that require a running server",
    "qt: marks tests that require a running qt application",
]


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[[tool.pydoc-markdown.loaders]]
type = "python"
search_path = ["herre"]

[tool.pydoc-markdown.renderer]
type = "docusaurus"
docs_base_path = "website/docs"


[[tool.mypy.overrides]]
module = ["oauthlib.*", "koil.*", "fakts.*"]
ignore_missing_imports = true
